ac_circ_buf: Fix some potential unsigned integer overflows
LGTM[0] pointed out some issues in ac_circ_buf.c regarding the
malloc(3)'s

  'Multiplication result may overflow 'unsigned int' before it is
   converted to 'size_t'.'

This is unlikely to hit in real life, but lets fix it anyway by casting
the size/count part of the calculation in the mallocs to size_t.

[0]: https://lgtm.com/

+++ src/ac_circ_buf.c
@@ -106,7 +106,7 @@ ac_circ_buf_t *ac_circ_buf_new(u32 size, u32 elem_sz)
 	} else {
 		cbuf->elem_sz = elem_sz;
 		cbuf->type = CPY_BUF;
-		cbuf->buf.cpy_buf = malloc(size * elem_sz);
+		cbuf->buf.cpy_buf = malloc((size_t)size * elem_sz);
 	}
 
 	return cbuf;
@@ -151,7 +151,7 @@ int ac_circ_buf_pushm(ac_circ_buf_t *cbuf, const void *buf, u32 count)
 		       count * sizeof(void *));
 	else
 		memcpy(cbuf->buf.cpy_buf + cbuf->head, buf,
-		       count * cbuf->elem_sz);
+		       (size_t)count * cbuf->elem_sz);
 
 	cbuf->head = (cbuf->head + (count * cbuf->elem_sz)) &
 		     ((cbuf->size - 1) * cbuf->elem_sz);
@@ -206,7 +206,7 @@ int ac_circ_buf_popm(ac_circ_buf_t *cbuf, void *buf, u32 count)
 		       count * sizeof(void *));
 	else
 		memcpy(buf, cbuf->buf.cpy_buf + cbuf->tail,
-		       count * cbuf->elem_sz);
+		       (size_t)count * cbuf->elem_sz);
 
 	cbuf->tail = (cbuf->tail + (count * cbuf->elem_sz)) &
 		     ((cbuf->size - 1) * cbuf->elem_sz);
